chore: refactor compliance_summary_service to align with new architecture
